fix: fetch_proton_paths on windows
